app/requirements.txt
albumentations==1.4.4
annotated-types==0.6.0
blinker==1.7.0
click==8.1.7
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.0
fsspec==2024.3.1
imageio==2.34.1
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
PyWavelets==1.4.1
PyYAML==6.0.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sympy==1.12
threadpoolctl==3.4.0
tifffile==2023.7.10
torch==2.3.0
torchvision==0.18.0
triton==2.3.0
typing_extensions==4.11.0
Werkzeug==3.0.2
zipp==3.18.1
